backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
psutil==5.9.6
pyautogui==0.9.54
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
schedule==1.2.0
watchdog==3.0.0
pynput==1.7.6
plyer==2.1.0
google-generativeai==0.8.0
python-pptx==0.6.21
reportlab==4.0.4
python-daemon==3.0.1
